move SelectedIndices to the index module
