+++ fast_forward/src/lib.rs
@@ -25,167 +25,10 @@
 //! To Find the `Key`: "Jon" with the `operation equals` is only one step necessary.
 //!
 
-use std::{
-    borrow::Cow,
-    cmp::{min, Ordering::*},
-    ops::{BitAnd, BitOr, Index},
-    slice,
-};
-
 pub mod collections;
 pub mod index;
 
 pub use collections::{Iter, ListIndexFilter};
-
-#[derive(Debug, Default, Clone)]
-#[repr(transparent)]
-pub struct SelectedIndices<'i>(Cow<'i, [usize]>);
-
-/// `SelIdx` (Selected Indices) is the result from quering (filter) a list.
-impl<'i> SelectedIndices<'i> {
-    #[inline]
-    pub fn new(i: usize) -> Self {
-        Self(Cow::Owned(vec![i]))
-    }
-
-    pub const fn empty() -> Self {
-        Self(Cow::Owned(Vec::new()))
-    }
-
-    pub const fn borrowed(s: &'i [usize]) -> Self {
-        Self(Cow::Borrowed(s))
-    }
-
-    pub const fn owned(v: Vec<usize>) -> Self {
-        Self(Cow::Owned(v))
-    }
-
-    #[inline]
-    pub fn iter(&'i self) -> slice::Iter<'i, usize> {
-        self.0.iter()
-    }
-
-    #[inline]
-    pub fn len(&self) -> usize {
-        self.0.len()
-    }
-
-    #[inline]
-    pub fn is_empty(&self) -> bool {
-        self.0.is_empty()
-    }
-}
-
-impl<'i> Index<usize> for SelectedIndices<'i> {
-    type Output = usize;
-
-    fn index(&self, index: usize) -> &Self::Output {
-        &self.0[index]
-    }
-}
-
-impl<'i, const N: usize> PartialEq<[usize; N]> for SelectedIndices<'i> {
-    fn eq(&self, other: &[usize; N]) -> bool {
-        (*self.0).eq(other)
-    }
-}
-
-impl<'i, const N: usize> PartialEq<SelectedIndices<'i>> for [usize; N] {
-    fn eq(&self, other: &SelectedIndices) -> bool {
-        (self).eq(&*other.0)
-    }
-}
-
-impl<'i> PartialEq for SelectedIndices<'i> {
-    fn eq(&self, other: &Self) -> bool {
-        self.0 == other.0
-    }
-}
-
-impl<'i> BitOr for SelectedIndices<'i> {
-    type Output = Self;
-
-    fn bitor(self, other: Self) -> Self::Output {
-        let lhs = &self.0;
-        let rhs = &other.0;
-
-        match (lhs.is_empty(), rhs.is_empty()) {
-            (false, false) => {
-                let (ll, lr) = (lhs.len(), rhs.len());
-                let mut v = Vec::with_capacity(ll + lr);
-
-                let (mut li, mut ri) = (0, 0);
-
-                loop {
-                    let (l, r) = (lhs[li], rhs[ri]);
-
-                    match l.cmp(&r) {
-                        Equal => {
-                            v.push(l);
-                            li += 1;
-                            ri += 1;
-                        }
-                        Less => {
-                            v.push(l);
-                            li += 1;
-                        }
-                        Greater => {
-                            v.push(r);
-                            ri += 1;
-                        }
-                    }
-
-                    if ll == li {
-                        v.extend(rhs[ri..].iter());
-                        return SelectedIndices::owned(v);
-                    } else if lr == ri {
-                        v.extend(lhs[li..].iter());
-                        return SelectedIndices::owned(v);
-                    }
-                }
-            }
-            (true, false) => other,
-            (false, true) => self,
-            (true, true) => SelectedIndices::empty(),
-        }
-    }
-}
-
-impl<'i> BitAnd for SelectedIndices<'i> {
-    type Output = Self;
-
-    fn bitand(self, other: Self) -> Self::Output {
-        let lhs = &self.0;
-        let rhs = &other.0;
-
-        if lhs.is_empty() || rhs.is_empty() {
-            return SelectedIndices::empty();
-        }
-
-        let (ll, lr) = (lhs.len(), rhs.len());
-        let mut v = Vec::with_capacity(min(ll, lr));
-
-        let (mut li, mut ri) = (0, 0);
-
-        loop {
-            let l = lhs[li];
-
-            match l.cmp(&rhs[ri]) {
-                Equal => {
-                    v.push(l);
-                    li += 1;
-                    ri += 1;
-                }
-                Less => li += 1,
-                Greater => ri += 1,
-            }
-
-            if li == ll || ri == lr {
-                return SelectedIndices::owned(v);
-            }
-        }
-    }
-}
 
 /// This `macro` is not a solution, it is more an POC (proof of concept)!
 /// The Problem with this macro is the visibility. This means, it can not hide internal fields,
@@ -295,7 +138,7 @@ macro_rules! fast {
             /// Panics if the positions are out of bound.
             ///
             #[allow(dead_code)]
-            fn filter<'i>(&'i self, filter: $crate::SelectedIndices<'i>) -> $crate::Iter<'i, $crate::collections::list::List<$item>> {
+            fn filter<'i>(&'i self, filter: $crate::index::SelectedIndices<'i>) -> $crate::Iter<'i, $crate::collections::list::List<$item>> {
                 use $crate::ListIndexFilter;
 
                 self._items_.filter(filter)
@@ -330,272 +173,7 @@ mod tests {
     use crate::{
         fast,
         index::{map::MapIndex, uint::UIntIndex, Retriever},
-        SelectedIndices,
     };
-    use std::borrow::Cow;
-
-    impl<'i> SelectedIndices<'i> {
-        fn from_slice(s: &'i [usize]) -> Self {
-            Self(Cow::Borrowed(s))
-        }
-    }
-
-    mod indices_or {
-        use super::*;
-
-        #[test]
-        fn both_empty() {
-            assert_eq!(
-                SelectedIndices::empty(),
-                SelectedIndices::empty() | SelectedIndices::empty()
-            );
-        }
-
-        #[test]
-        fn only_left() {
-            assert_eq!(
-                SelectedIndices::from_slice(&[1, 2]),
-                SelectedIndices::from_slice(&[1, 2]) | SelectedIndices::empty()
-            );
-        }
-
-        #[test]
-        fn only_right() {
-            assert_eq!(
-                SelectedIndices::from_slice(&[1, 2]),
-                SelectedIndices::empty() | SelectedIndices::from_slice(&[1, 2])
-            );
-        }
-
-        #[test]
-        fn diff_len() {
-            assert_eq!(
-                SelectedIndices::from_slice(&[1, 2, 3]),
-                SelectedIndices::new(1) | SelectedIndices::from_slice(&[2, 3]),
-            );
-            assert_eq!(
-                SelectedIndices::from_slice(&[1, 2, 3]),
-                SelectedIndices::from_slice(&[2, 3]) | SelectedIndices::new(1)
-            );
-        }
-
-        #[test]
-        fn overlapping_simple() {
-            assert_eq!(
-                SelectedIndices::from_slice(&[1, 2, 3]),
-                SelectedIndices::from_slice(&[1, 2]) | SelectedIndices::from_slice(&[2, 3])
-            );
-            assert_eq!(
-                SelectedIndices::from_slice(&[1, 2, 3]),
-                SelectedIndices::from_slice(&[2, 3]) | SelectedIndices::from_slice(&[1, 2])
-            );
-        }
-
-        #[test]
-        fn overlapping_diff_len() {
-            // 1, 2, 8, 9, 12
-            // 2, 5, 6, 10
-            assert_eq!(
-                SelectedIndices::from_slice(&[1, 2, 5, 6, 8, 9, 10, 12]),
-                SelectedIndices::from_slice(&[1, 2, 8, 9, 12])
-                    | SelectedIndices::from_slice(&[2, 5, 6, 10])
-            );
-
-            // 2, 5, 6, 10
-            // 1, 2, 8, 9, 12
-            assert_eq!(
-                SelectedIndices::from_slice(&[1, 2, 5, 6, 8, 9, 10, 12]),
-                SelectedIndices::from_slice(&[2, 5, 6, 10])
-                    | SelectedIndices::from_slice(&[1, 2, 8, 9, 12])
-            );
-        }
-    }
-
-    mod indices_and {
-        use super::*;
-
-        #[test]
-        fn both_empty() {
-            assert_eq!(
-                SelectedIndices::empty(),
-                SelectedIndices::empty() & SelectedIndices::empty()
-            );
-        }
-
-        #[test]
-        fn only_left() {
-            assert_eq!(
-                SelectedIndices::empty(),
-                SelectedIndices::from_slice(&[1, 2]) & SelectedIndices::empty()
-            );
-        }
-
-        #[test]
-        fn only_right() {
-            assert_eq!(
-                SelectedIndices::empty(),
-                SelectedIndices::empty() & SelectedIndices::from_slice(&[1, 2])
-            );
-        }
-
-        #[test]
-        fn diff_len() {
-            assert_eq!(
-                SelectedIndices::empty(),
-                SelectedIndices::from_slice(&[1]) & SelectedIndices::from_slice(&[2, 3])
-            );
-            assert_eq!(
-                SelectedIndices::empty(),
-                SelectedIndices::from_slice(&[2, 3]) & SelectedIndices::from_slice(&[1])
-            );
-
-            assert_eq!(
-                [2],
-                SelectedIndices::from_slice(&[2]) & SelectedIndices::from_slice(&[2, 5])
-            );
-            assert_eq!(
-                [2],
-                SelectedIndices::from_slice(&[2]) & SelectedIndices::from_slice(&[1, 2, 3])
-            );
-            assert_eq!(
-                [2],
-                SelectedIndices::from_slice(&[2]) & SelectedIndices::from_slice(&[0, 1, 2])
-            );
-
-            assert_eq!(
-                [2],
-                SelectedIndices::from_slice(&[2, 5]) & SelectedIndices::from_slice(&[2])
-            );
-            assert_eq!(
-                [2],
-                SelectedIndices::from_slice(&[1, 2, 3]) & SelectedIndices::from_slice(&[2])
-            );
-            assert_eq!(
-                [2],
-                SelectedIndices::from_slice(&[0, 1, 2]) & SelectedIndices::from_slice(&[2])
-            );
-        }
-
-        #[test]
-        fn overlapping_simple() {
-            assert_eq!(
-                [2],
-                SelectedIndices::from_slice(&[1, 2]) & SelectedIndices::from_slice(&[2, 3]),
-            );
-            assert_eq!(
-                [2],
-                SelectedIndices::from_slice(&[2, 3]) & SelectedIndices::from_slice(&[1, 2]),
-            );
-
-            assert_eq!(
-                [1],
-                SelectedIndices::from_slice(&[1, 2]) & SelectedIndices::from_slice(&[1, 3]),
-            );
-            assert_eq!(
-                [1],
-                SelectedIndices::from_slice(&[1, 3]) & SelectedIndices::from_slice(&[1, 2]),
-            );
-        }
-
-        #[test]
-        fn overlapping_diff_len() {
-            // 1, 2, 8, 9, 12
-            // 2, 5, 6, 10
-            assert_eq!(
-                [2, 12],
-                SelectedIndices::from_slice(&[1, 2, 8, 9, 12])
-                    & SelectedIndices::from_slice(&[2, 5, 6, 10, 12, 13, 15])
-            );
-
-            // 2, 5, 6, 10
-            // 1, 2, 8, 9, 12
-            assert_eq!(
-                [2, 12],
-                SelectedIndices::from_slice(&[2, 5, 6, 10, 12, 13, 15])
-                    & SelectedIndices::from_slice(&[1, 2, 8, 9, 12])
-            );
-        }
-    }
-
-    mod query {
-        use super::*;
-
-        struct List(Vec<usize>);
-
-        impl List {
-            fn eq(&self, i: usize) -> SelectedIndices<'_> {
-                match self.0.binary_search(&i) {
-                    Ok(pos) => SelectedIndices::owned(vec![pos]),
-                    Err(_) => SelectedIndices::empty(),
-                }
-            }
-        }
-
-        fn values() -> List {
-            List(vec![0, 1, 2, 3])
-        }
-
-        #[test]
-        fn filter() {
-            let l = values();
-            assert_eq!(1, l.eq(1)[0]);
-            assert_eq!(SelectedIndices::empty(), values().eq(99));
-        }
-
-        #[test]
-        fn and() {
-            let l = values();
-            assert_eq!(1, (l.eq(1) & l.eq(1))[0]);
-            assert_eq!(SelectedIndices::empty(), (l.eq(1) & l.eq(2)));
-        }
-
-        #[test]
-        fn or() {
-            let l = values();
-            assert_eq!([1, 2], l.eq(1) | l.eq(2));
-            assert_eq!([1], l.eq(1) | l.eq(99));
-            assert_eq!([1], l.eq(99) | l.eq(1));
-        }
-
-        #[test]
-        fn and_or() {
-            let l = values();
-            // (1 and 1) or 2 => [1, 2]
-            assert_eq!([1, 2], l.eq(1) & l.eq(1) | l.eq(2));
-            // (1 and 2) or 3 => [3]
-            assert_eq!([3], l.eq(1) & l.eq(2) | l.eq(3));
-        }
-
-        #[test]
-        fn or_and_12() {
-            let l = values();
-            // 1 or (2 and 2) => [1, 2]
-            assert_eq!([1, 2], l.eq(1) | l.eq(2) & l.eq(2));
-            // 1 or (3 and 2) => [1]
-            assert_eq!([1], l.eq(1) | l.eq(3) & l.eq(2));
-        }
-
-        #[test]
-        fn or_and_3() {
-            let l = values();
-            // 3 or (2 and 1) => [3]
-            assert_eq!([3], l.eq(3) | l.eq(2) & l.eq(1));
-        }
-
-        #[test]
-        fn and_or_and_2() {
-            let l = values();
-            // (2 and 2) or (2 and 1) => [2]
-            assert_eq!([2], l.eq(2) & l.eq(2) | l.eq(2) & l.eq(1));
-        }
-
-        #[test]
-        fn and_or_and_03() {
-            let l = values();
-            // 0 or (1 and 2) or 3) => [0, 3]
-            assert_eq!([0, 3], l.eq(0) | l.eq(1) & l.eq(2) | l.eq(3));
-        }
-    }
 
     #[derive(Debug, Eq, PartialEq, Clone)]
     struct Car(usize, String);
